requirements.txt
gradio
transformers
torch
streamlit
requests
python-dotenv
